fix: role column user에 통합
