+++ helper-core/src/main/java/org/example/domain/teacher/model/Teacher.java
@@ -15,6 +15,4 @@ public class Teacher {
     private final String userId;
 
     private final String name;
-
-    private final Role role;
 }
