+++ helper-infrastructure/src/main/java/org/example/global/security/auth/CustomUserDetail.java
@@ -1,10 +1,24 @@
 package org.example.global.security.auth;
 
 import org.example.domain.auth.model.Role;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
 
-public interface CustomUserDetail {
+import java.util.Collection;
 
-    String getUserId();
+public record CustomUserDetail (String userId, Role role) implements UserDetails {
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return null;
+    }
 
-    Role getRole();
+    @Override
+    public String getPassword() {
+        return null;
+    }
+
+    @Override
+    public String getUsername() {
+        return null;
+    }
 }
